requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
rich>=13.7.0
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
pyfiglet>=1.0.2
colorama>=0.4.6
emojis>=0.7.0
tiktoken>=0.5.0

